% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plotting the observations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing a state in the USA.}

\item{year}{An integer meaning the year of observation.}
}
\value{
The function plots the map of the american state chosen by the user.
Black dots represent the place of observed fatal accidents.
If user define an invalid state number or year for which data are not available,
the function returns an error. If in a given state/ year pair there are no observed accidents,
the function returns the message "no  accidents to plot".
}
\description{
The "fars_map_state" function plots the observations of fatal accidents on the map of the
provided state in the year defined by the user.
}
\examples{
fars_map_state(39,2013)

}

