% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading data from the Fatality Analysis Reporting System
The fars_read function reads data file into a dplyr data table.
If the data file does not exist in the working directory, this function throws an error.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving a filename.}
}
\value{
The function reads the given file and creates a dplyr data table from it.
If file with the given name does not exist, the function returns an error message.
}
\description{
Reading data from the Fatality Analysis Reporting System
The fars_read function reads data file into a dplyr data table.
If the data file does not exist in the working directory, this function throws an error.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}

}

