% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generating a list of tables}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of integers representing years.}
}
\value{
A list with the same number of data tables as the number of years provided by the user
if the data of those years are available. For the years not represented,
the function sends warning message(s).
}
\description{
The "fars_read_years" function creates a list of dplyr data tables
where each data table represents a year provided by the user,
every data tables contain two columns (month, year)
and the number of rows is defined by the number of observations.
}
\examples{
fars_read_years(2013)
fars_read_years(2013:2015)

}

