% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generating tables showing observations by months and years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of integers representing years.}
}
\value{
The function creates a tidyr data table, in which columns represent years and
rows show the months of observations.
For the years not represented, the function sends warning message(s).
}
\description{
The "fars_summarize_years" function summarizes the number of observations by months for the years provided by the user.
The results are presented in a tidyr data table.
}
\examples{
fars_summarize_years(2013:2015)

}

